+++ .prettierrc.js
@@ -1,9 +1,1 @@
-module.exports = {
-    printWidth: 100,
-    singleQuote: true,
-    quoteProps: 'as-needed',
-    trailingComma: 'all',
-    tabWidth: 4,
-    useTabs: false,
-    semi: true,
-};
+module.exports = require('apteryx-prettier-config');