+++ src/errors/Messages.ts
@@ -9,8 +9,12 @@ const Messages = {
 
     COMMAND_MISSING_METHOD: (name: string, method: string) =>
         `Command ${name} is missing its ${method} method.`,
+    COMPONENT_MISSING_METHOD: (name: string, method: string) =>
+        `Component ${name} is missing its ${method} method.`,
     EVENT_MISSING_METHOD: (name: string, method: string) =>
         `Event ${name} is missing its ${method} method.`,
+
+    COMMAND_NOT_IMPLEMENTED: (name: string) => `Command ${name} has not been implemented yet.`,
 };
 
 for (const [key, val] of Object.entries(Messages)) {