+++ src/index.ts
@@ -12,15 +12,20 @@ export * from './structures/Args';
 export * from './structures/Base';
 export * from './structures/Client';
 export * from './structures/Command';
+export * from './structures/Component';
 export * from './structures/Event';
 export * from './structures/Plugin';
 export * from './structures/Precondition';
 export * as Preconditions from './preconditions/index';
 
 // Managers
-export type { CommandManager } from './managers/CommandManager';
-export type { EventManager } from './managers/EventManager';
-export type { PluginManager } from './managers/PluginManager';
+export * from './managers/BaseManager';
+export * from './managers/CommandManager';
+export * from './managers/ComponentManager';
+export * from './managers/EventManager';
+export * from './managers/MapManager';
+export * from './managers/PluginManager';
+export * from './managers/SetManager';
 
 // Errors
 export * from './errors';
@@ -28,6 +33,5 @@ export * from './errors';
 // Types
 export { Events } from './types/Events';
 
-// Util
-export * from './utils/CustomId';
+// Utils
 export * as Regexes from './utils/Regexes';